refactor(checkmarx): remove example app and backend from workspace
