--- workspaces/checkmarx/packages/app/.eslintignore
@@ -1,1 +1,0 @@
-public
