--- workspaces/checkmarx/packages/app/CHANGELOG.md
@@ -1,7 +1,0 @@
-# app
-
-## 0.0.1
-
-### Patch Changes
-
-- Initial local development app scaffold for the Checkmarx workspace.
