--- workspaces/checkmarx/packages/app/knip-report.md
@@ -1,21 +1,0 @@
-# Knip report
-
-## Unused dependencies (5)
-
-| Name                            | Location     | Severity |
-| :------------------------------ | :----------- | :------- |
-| @backstage/plugin-catalog-react | package.json | error    |
-| @backstage/catalog-model        | package.json | error    |
-| react-router                    | package.json | error    |
-| react-use                       | package.json | error    |
-| history                         | package.json | error    |
-
-## Unused devDependencies (5)
-
-| Name                        | Location     | Severity |
-| :-------------------------- | :----------- | :------- |
-| @testing-library/user-event | package.json | error    |
-| @backstage/test-utils       | package.json | error    |
-| @testing-library/dom        | package.json | error    |
-| @playwright/test            | package.json | error    |
-| cross-env                   | package.json | error    |
