--- workspaces/checkmarx/packages/backend/CHANGELOG.md
@@ -1,7 +1,0 @@
-# backend
-
-## 0.0.1
-
-### Patch Changes
-
-- Initial local development backend scaffold for the Checkmarx workspace.
