--- workspaces/checkmarx/packages/backend/knip-report.md
@@ -1,25 +1,0 @@
-# Knip report
-
-## Unused dependencies (10)
-
-| Name                                                  | Location     | Severity |
-| :---------------------------------------------------- | :----------- | :------- |
-| @backstage/plugin-auth-backend-module-github-provider | package.json | error    |
-| @backstage/backend-plugin-api                         | package.json | error    |
-| @backstage/plugin-auth-node                           | package.json | error    |
-| @backstage/config                                     | package.json | error    |
-| better-sqlite3                                        | package.json | error    |
-| dockerode                                             | package.json | error    |
-| node-gyp                                              | package.json | error    |
-| winston                                               | package.json | error    |
-| app                                                   | package.json | error    |
-| pg                                                    | package.json | error    |
-
-## Unused devDependencies (4)
-
-| Name                             | Location     | Severity |
-| :------------------------------- | :----------- | :------- |
-| @types/express-serve-static-core | package.json | error    |
-| @types/dockerode                 | package.json | error    |
-| @types/express                   | package.json | error    |
-| @types/luxon                     | package.json | error    |
