--- workspaces/checkmarx/packages/backend/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "backend",
-  "version": "0.0.1",
-  "main": "dist/index.cjs.js",
-  "types": "src/index.ts",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/checkmarx/packages/backend"
-  },
-  "backstage": {
-    "role": "backend"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test",
-    "clean": "backstage-cli package clean",
-    "build-image": "docker build ../.. -f Dockerfile --tag backstage"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-checkmarx-backend": "workspace:^",
-    "@backstage/backend-defaults": "backstage:^",
-    "@backstage/backend-plugin-api": "backstage:^",
-    "@backstage/config": "backstage:^",
-    "@backstage/plugin-app-backend": "backstage:^",
-    "@backstage/plugin-auth-backend": "backstage:^",
-    "@backstage/plugin-auth-backend-module-github-provider": "backstage:^",
-    "@backstage/plugin-auth-backend-module-guest-provider": "backstage:^",
-    "@backstage/plugin-auth-node": "backstage:^",
-    "@backstage/plugin-catalog-backend": "backstage:^",
-    "@backstage/plugin-proxy-backend": "backstage:^",
-    "app": "link:../app",
-    "better-sqlite3": "^12.0.0",
-    "node-gyp": "^9.0.0",
-    "pg": "^8.11.3",
-    "winston": "^3.2.1"
-  },
-  "devDependencies": {
-    "@backstage/cli": "backstage:^",
-    "@types/express": "^4.17.6",
-    "@types/express-serve-static-core": "^5.0.0",
-    "@types/luxon": "^3.0.0"
-  },
-  "files": [
-    "dist"
-  ]
-}
