+++ workspaces/checkmarx/plugins/checkmarx-backend/README.md
@@ -11,10 +11,10 @@ This backend plugin resolves Checkmarx One data for a Backstage catalog entity b
 ## Installation
 
 ```bash
-yarn --cwd packages/backend add @backstage-community/plugin-checkmarx-backend
+yarn add @backstage-community/plugin-checkmarx-backend
 ```
 
-Add the plugin to `packages/backend/src/index.ts`:
+Add the plugin to your Backstage backend:
 
 ```ts
 backend.add(import('@backstage-community/plugin-checkmarx-backend'));
