+++ workspaces/checkmarx/plugins/checkmarx-backend/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@backstage-community/plugin-checkmarx-backend",
   "version": "0.0.0",
+  "description": "Backstage backend plugin for retrieving Checkmarx scan data for catalog entities",
   "backstage": {
     "role": "backend-plugin",
     "pluginId": "checkmarx",
@@ -8,7 +9,10 @@
       "@backstage-community/plugin-checkmarx",
       "@backstage-community/plugin-checkmarx-backend",
       "@backstage-community/plugin-checkmarx-react"
-    ]
+    ],
+    "features": {
+      ".": "@backstage/BackendFeature"
+    }
   },
   "publishConfig": {
     "access": "public",
@@ -22,8 +26,8 @@
     "directory": "workspaces/checkmarx/plugins/checkmarx-backend"
   },
   "license": "Apache-2.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "dist/index.cjs.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist",
     "config.d.ts"
@@ -55,11 +59,19 @@
     "@backstage/cli": "backstage:^",
     "@backstage/plugin-auth-backend": "backstage:^",
     "@backstage/plugin-auth-backend-module-guest-provider": "backstage:^",
+    "@backstage/plugin-catalog-backend": "backstage:^",
     "@types/express": "^4.17.21",
     "@types/node-fetch": "^2.5.12",
     "@types/supertest": "^7.0.0",
     "msw": "^1.0.0",
     "supertest": "^7.0.0"
   },
-  "configSchema": "config.d.ts"
+  "configSchema": "config.d.ts",
+  "typesVersions": {
+    "*": {
+      "package.json": [
+        "package.json"
+      ]
+    }
+  }
 }
