+++ workspaces/checkmarx/plugins/checkmarx-react/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@backstage-community/plugin-checkmarx-react",
   "version": "0.0.0",
+  "description": "Shared types and entity helpers for the Backstage Checkmarx plugins",
   "backstage": {
     "role": "common-library",
     "pluginId": "checkmarx",
