+++ workspaces/checkmarx/tsconfig.json
@@ -1,11 +1,6 @@
 {
   "extends": "@backstage/cli/config/tsconfig.json",
-  "include": [
-    "packages/*/src",
-    "plugins/*/src",
-    "plugins/*/dev",
-    "plugins/*/migrations"
-  ],
+  "include": ["plugins/*/src", "plugins/*/dev", "plugins/*/migrations"],
   "exclude": ["node_modules"],
   "compilerOptions": {
     "outDir": "dist-types",
